refactor flash, downloaded and worklist
