+++ src/mpflash/tests/test_sort.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from mpflash.flash import find_firmware, local_firmwares
+from mpflash.flash import downloaded_firmwares, find_downloaded_firmware
 
 pytestmark = [pytest.mark.mpflash]
 ####################
@@ -12,13 +12,13 @@ pytestmark = [pytest.mark.mpflash]
 
 def test_load(tmp_path: Path):
     test_fw_folder = Path(__file__).parent / "data" / "firmware"
-    all = local_firmwares(test_fw_folder)
+    all = downloaded_firmwares(test_fw_folder)
     assert len(all) > 0
 
 
 def test_find():
     test_fw_folder = Path(__file__).parent / "data" / "firmware"
-    fws = find_firmware(fw_folder=test_fw_folder, port="samd", board="SEEED_WIO_TERMINAL",version="preview")
+    fws = find_downloaded_firmware(fw_folder=test_fw_folder, port="samd", board="SEEED_WIO_TERMINAL", version="preview")
     assert len(fws) > 0
     assert fws[0]["board"] == "SEEED_WIO_TERMINAL"
     assert int(fws[-1]["build"]) > 0
